backend/requirements.txt
psycopg2-binary==2.9.9
Flask-Limiter>=3.5.0
gunicorn==22.0.0
python-dotenv==1.0.1
google-cloud-vision==3.7.4
Flask-SQLAlchemy==3.1.1
google-cloud-storage==2.18.2
requests==2.32.3
Flask-Migrate==4.0.7
flask-cors>=4.0.0
Flask==3.0.3
google-genai>=1.38.0